Abstract controller added, main game controllers changed
