+++ src/main/java/com/example/controller/ControllerChoice.java
@@ -1,11 +1,11 @@
 package com.example.controller;
 
+import com.example.controller.logic.ControllerAI;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-
 import javafx.stage.Stage;
 
 import java.io.IOException;
