+++ src/main/java/com/example/controller/ControllerMain.java
@@ -1,5 +1,6 @@
 package com.example.controller;
 
+import com.example.controller.logic.ControllerTwoPlayer;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
