--- src/main/java/com/example/controller/ControllerTwoPlayer.java
@@ -1,126 +1,0 @@
-package com.example.controller;
-
-import com.example.style.Style;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.ResourceBundle;
-
-public class ControllerTwoPlayer implements Initializable {
-    @FXML
-    private Button restart;
-    @FXML
-    private Button button1, button2, button3, button4, button5, button6, button7, button8, button9;
-    @FXML
-    private Label message;
-
-    private int playerTurn = 0;
-    private ArrayList<Button> buttons;
-
-    private Style style = new Style();
-
-    @Override
-    public void initialize(URL url, ResourceBundle resourceBundle) {
-        buttons = new ArrayList<>(Arrays.asList(button1, button2, button3, button4,
-                        button5, button6, button7, button8, button9));
-        restart.getStyleClass().add("hover");
-        buttons.forEach(button -> {
-            style.setButtonHoverDisabled(button);
-            setupButton(button);
-            button.setFocusTraversable(false);
-        });
-        newGame();
-    }
-
-
-    public void newGame() {
-        restart.setOnMouseClicked(mouseEvent -> {
-        buttons.forEach(this::resetButton);
-        message.setText("Tic-Tac-Toe");
-        });
-    }
-
-    public void resetButton (Button button) {
-        button.setText("");
-        button.setDisable(false);
-        style.resetChoice(button);
-    }
-
-    private void setupButton(Button button) {
-        button.setOnMouseClicked(mouseEvent -> {
-            setPlayerSymbol(button);
-            style.setButtonStyle(button, button.getText());
-            button.setDisable(true);
-            checkIfGameIsOver();
-        });
-    }
-
-    public void setPlayerSymbol(Button button){
-        if (playerTurn % 2 == 0) {
-            button.setText("X");
-            playerTurn = 1;
-        } else {
-            button.setText("O");
-            playerTurn = 0;
-        }
-    }
-
-    private void checkIfGameIsOver() {
-        for (int a = 0; a < 8; a++) {
-            String line = switch (a) {
-                case 0 -> button1.getText() + button2.getText() + button3.getText();
-                case 1 -> button4.getText() + button5.getText() + button6.getText();
-                case 2 -> button7.getText() + button8.getText() + button9.getText();
-                case 3 -> button1.getText() + button5.getText() + button9.getText();
-                case 4 -> button3.getText() + button5.getText() + button7.getText();
-                case 5 -> button1.getText() + button4.getText() + button7.getText();
-                case 6 -> button2.getText() + button5.getText() + button8.getText();
-                case 7 -> button3.getText() + button6.getText() + button9.getText();
-                default -> null;
-            };
-
-            checkLine(line);
-        }
-    }
-
-    private void checkLine(String line) {
-        if (line.equals("XXX")) {
-            gameOver("Win X!");
-        } else if (line.equals("OOO")) {
-            gameOver("Win O!");
-        }
-    }
-
-    private void gameOver(String messageText) {
-        message.setText(messageText);
-        buttons.forEach(button -> button.setDisable(true));
-    }
-//    public void checkIfGameIsOver(){
-//        for (int i = 0; i < 8; i++) {
-//            String line = switch (i) {
-//                case 0 -> button1.getText() + button2.getText() + button3.getText();
-//                case 1 -> button4.getText() + button5.getText() + button6.getText();
-//                case 2 -> button9.getText() + button8.getText() + button9.getText();
-//                case 3 -> button1.getText() + button5.getText() + button9.getText();
-//                case 4 -> button3.getText() + button5.getText() + button7.getText();
-//                case 5 -> button1.getText() + button4.getText() + button7.getText();
-//                case 6 -> button2.getText() + button5.getText() + button8.getText();
-//                case 7 -> button3.getText() + button6.getText() + button9.getText();
-//                default -> null;
-//            };
-//
-//            if (line.equals("XXX")) {
-//                message.setText("Win X");
-//                buttons.forEach(button -> button.setDisable(true));
-//            } else if (line.equals("OOO")) {
-//                message.setText("Win O");
-//                buttons.forEach(button -> button.setDisable(true));
-//            }
-//        }
-//    }
-}
